+++ asf_search/export/csv.py
@@ -62,8 +62,8 @@ fieldnames = (
     "absoluteBurstID",
     "fullBurstID",
     "burstIndex",
-    "burstAnxTime",
-    "timeFromAnxSeconds",
+    "azimuthTime",
+    "azimuthAnxTime",
     "samplesPerBurst",
     "subswath"
 )
@@ -155,8 +155,8 @@ class CSVStreamArray(list):
             "absoluteBurstID":  p['burst']['absoluteBurstID'] if p['processingLevel'] == 'BURST' else None,
             "fullBurstID":  p['burst']['fullBurstID'] if p['processingLevel'] == 'BURST' else None,
             "burstIndex":   p['burst']['burstIndex'] if p['processingLevel'] == 'BURST' else None,
-            "burstAnxTime": p['burst']['burstAnxTime'] if p['processingLevel'] == 'BURST' else None,
-            "timeFromAnxSeconds":   p['burst']['timeFromAnxSeconds'] if p['processingLevel'] == 'BURST' else None,
+            "azimuthTime": p['burst']['azimuthTime'] if p['processingLevel'] == 'BURST' else None,
+            "azimuthAnxTime": p['burst']['azimuthAnxTime'] if p['processingLevel'] == 'BURST' else None,
             "samplesPerBurst":  p['burst']['samplesPerBurst'] if p['processingLevel'] == 'BURST' else None,
             "subswath": p['burst']['subswath'] if p['processingLevel'] == 'BURST' else None
         }